requirements.txt
pandas
lxml
html5lib
beautifulsoup4
pylint
pytest
dbt-core
dbt-snowflake
matplotlib
seaborn
